requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8
drf-spectacular>=0.15.1,<0.16